+++ src/components/Calendar/CalendarRange/CalendarRange.tsx
@@ -1,5 +1,5 @@
 import { Dayjs } from "dayjs";
-import { Card, Icon, DateRange, Slide } from "../..";
+import { Icon, DateRange, Slide } from "../..";
 import React, {
   forwardRef,
   memo,
@@ -7,11 +7,12 @@ import React, {
   useEffect,
   useState,
 } from "react";
-import { Container, IconContainer } from "../styled";
+import { Container, IconContainer, Card } from "../styled";
 import { Action, Actions } from "../internal/Actions";
 import { ClickState, ClickStateType } from "./constants";
 import { InnerCalendarRange } from "../internal/InnerCalendarRange/InnerCalendarRange";
 import { YearMonths } from "../internal/YearMonths";
+import { useTheme } from "../../../themes";
 
 export type CalendarRangeProps = React.HTMLAttributes<HTMLDivElement> & {
   /**
@@ -25,9 +26,26 @@ export type CalendarRangeProps = React.HTMLAttributes<HTMLDivElement> & {
    */
   endDate: Dayjs;
   /**
+   * カレンダーに表示する年月のフォーマット
+   */
+  monthFormat?: string;
+  /**
+   * カレンダーに表示する曜日のリスト
+   * @memo dayjs().format("ddd") で対応したいが、階層が深くなったりするので一旦静的な値で対処
+   */
+  weekList?: string[];
+  /**
+   * デフォルトで選択されているアクション
+   */
+  defaultClickAction?: string;
+  /**
    * カレンダーの左に表示するアクション
    */
   actions?: Action[];
+  /**
+   * アクションをクリックしたときの挙動
+   */
+  onClickAction?: (action: Action) => void;
   /**
    * 親コンポーネントで calendar を任意のタイミングで閉じたい場合に使用する
    */
@@ -59,7 +77,11 @@ export const CalendarRange = forwardRef<HTMLDivElement, CalendarRangeProps>(
     {
       startDate,
       endDate,
+      monthFormat = "YYYY年M月",
+      weekList = ["日", "月", "火", "水", "木", "金", "土"],
+      defaultClickAction,
       actions,
+      onClickAction,
       onClose,
       isOutsideRange = () => false,
       onClickCloseButton,
@@ -68,6 +90,7 @@ export const CalendarRange = forwardRef<HTMLDivElement, CalendarRangeProps>(
     },
     ref,
   ) {
+    const theme = useTheme();
     const [current, setCurrent] = React.useState<Dayjs>(startDate);
     const [yearIsOpen, setYearIsOpen] = React.useState(false);
 
@@ -111,14 +134,12 @@ export const CalendarRange = forwardRef<HTMLDivElement, CalendarRangeProps>(
     );
 
     return (
-      <Card
-        ref={ref}
-        display="flex"
-        width="fit-content"
-        style={{ overflow: "hidden" }}
-        {...rest}
-      >
-        <Actions actions={actions} />
+      <Card ref={ref} {...rest}>
+        <Actions
+          defaultClickAction={defaultClickAction}
+          actions={actions}
+          onClickAction={onClickAction}
+        />
         <Container>
           <div style={{ position: "relative", zIndex: 1 }}>
             <Slide unmountOnExit in={yearIsOpen} direction="up">
@@ -135,6 +156,8 @@ export const CalendarRange = forwardRef<HTMLDivElement, CalendarRangeProps>(
             <InnerCalendarRange
               startDate={startDate}
               endDate={endDate}
+              monthFormat={monthFormat}
+              weekList={weekList}
               current={current}
               yearIsOpen={yearIsOpen}
               isOutsideRange={isOutsideRange}
@@ -145,7 +168,7 @@ export const CalendarRange = forwardRef<HTMLDivElement, CalendarRangeProps>(
         </Container>
         {onClickCloseButton && (
           <IconContainer expanded={yearIsOpen} onClick={onClickCloseButton}>
-            <Icon name="close" />
+            <Icon name="close" color={theme.palette.black} />
           </IconContainer>
         )}
       </Card>