+++ src/components/Calendar/CalendarRange/internal/Day/styled.ts
@@ -17,8 +17,7 @@ export const DayStyle = styled.button`
 
   &:hover {
     transition: 0.3s;
-    background-color: ${({ theme }) => theme.palette.primary.main};
-    color: ${({ theme }) => theme.palette.white};
+    background-color: ${({ theme }) => theme.palette.primary.highlight};
   }
 `;
 
@@ -54,9 +53,16 @@ export const DayStart = styled(EdgeDate)`
   &::before {
     transform: translateX(100%);
   }
+  &:hover {
+    background-color: ${({ theme }) => theme.palette.primary.main};
+  }
 `;
 
-export const DayEnd = styled(EdgeDate)``;
+export const DayEnd = styled(EdgeDate)`
+  &:hover {
+    background-color: ${({ theme }) => theme.palette.primary.main};
+  }
+`;
 
 export const DayBetween = styled(SelectedDate)``;
 
@@ -64,6 +70,7 @@ export const DisableDayContainer = styled(DayStyle)`
   color: ${({ theme }) => theme.palette.gray.main};
   cursor: not-allowed;
   &:hover {
-    background-color: ${({ theme }) => theme.palette.gray.light};
+    background-color: transparent;
+    color: ${({ theme }) => theme.palette.gray.main};
   }
 `;