+++ src/components/Calendar/internal/Actions/Actions.tsx
@@ -9,36 +9,46 @@ export type Action = {
 };
 
 export const Actions = memo(function Actions({
+  defaultClickAction,
   actions,
+  onClickAction,
 }: {
+  defaultClickAction?: string;
   actions?: Action[];
+  onClickAction?: (action: Action) => void;
 }) {
   const theme = useTheme();
-  const [clickedAction, setClickedAction] = useState<string | null>(null);
+  const [clickedAction, setClickedAction] = useState<string | undefined>(
+    defaultClickAction,
+  );
 
   const handleClick = (action: Action) => () => {
     setClickedAction(action.text);
     action.onClick();
+    onClickAction?.(action);
   };
 
   return actions ? (
     <Flex display="flex">
-      <ActionsContainer>
-        {actions.map((action, i) => (
-          <Action
-            key={`${action.text}-${i.toString()}`}
-            clicked={clickedAction === action.text}
-            onClick={handleClick(action)}
-          >
-            {action.text}
-          </Action>
-        ))}
+      <ActionsContainer height="380px">
+        <>
+          {actions.map((action, i) => (
+            <Action
+              key={`${action.text}-${i.toString()}`}
+              clicked={clickedAction === action.text}
+              onClick={handleClick(action)}
+            >
+              {action.text}
+            </Action>
+          ))}
+        </>
       </ActionsContainer>
       <Divider
         ml={1}
         mr={1}
         orientation="vertical"
-        color={theme.palette.divider}
+        // MEMO: divider デザイントークンで basic[200] に該当するものがないので、一旦 gray で代用
+        color={theme.palette.gray.light}
       />
     </Flex>
   ) : null;