+++ src/components/Calendar/internal/InnerCalendar/InnerCalendar.tsx
@@ -3,7 +3,7 @@ import dayjs, { Dayjs } from "dayjs";
 import { Icon, ScrollArea } from "../../..";
 import { useTheme } from "styled-components";
 import { useScrollCalendar } from "../../hooks/useScrollCalendar";
-import { HEIGHT, weekList } from "../../constants";
+import { HEIGHT } from "../../constants";
 import {
   CalendarContainer,
   CalendarMonth,
@@ -16,6 +16,8 @@ import { Day } from "../../Calendar/internal/Day";
 
 type Props = {
   date: Dayjs;
+  monthFormat?: string;
+  weekList?: string[];
   current: Dayjs;
   yearIsOpen: boolean;
   onYearIsOpen: (yearIsOpen: boolean) => void;
@@ -24,6 +26,8 @@ type Props = {
 };
 export const InnerCalendar: React.FC<Props> = ({
   date,
+  weekList,
+  monthFormat,
   current,
   yearIsOpen,
   onYearIsOpen,
@@ -51,7 +55,7 @@ export const InnerCalendar: React.FC<Props> = ({
             {/* 年月の表示 */}
             <CalendarMonth expanded={yearIsOpen}>
               <TitleContainer expanded={yearIsOpen} weight="bold" size="xl">
-                {m.format("YYYY年MM月")}
+                {m.format(monthFormat)}
               </TitleContainer>
               <IconButton
                 expanded={yearIsOpen}
@@ -68,9 +72,7 @@ export const InnerCalendar: React.FC<Props> = ({
             {/* カレンダーの表示 */}
             <CalendarContainer>
               {/* 曜日の表示 */}
-              {weekList["ja"].map((week) => (
-                <DayStyle key={week}>{week}</DayStyle>
-              ))}
+              {weekList?.map((week) => <DayStyle key={week}>{week}</DayStyle>)}
 
               {/* 開始曜日まで空白をセット */}
               {Array.from(new Array(m.startOf("month").day()), (_, i) => (