+++ src/components/DateField/DateField/styled.ts
@@ -6,7 +6,7 @@ export const InputContainer = styled.div<{
 }>`
   display: flex;
   align-items: center;
-  width: 150px;
+  width: fit-content;
   /* MEMO: for calendar icon */
   padding: 0 8px 0 0;
   border: 0;
@@ -22,7 +22,6 @@ export const InputContainer = styled.div<{
     ${({ theme, error }) =>
       error ? theme.palette.danger.main : theme.palette.divider};
   border-radius: ${({ theme }) => theme.radius}px;
-  overflow: scroll;
   /* MEMO: To take a place that display LastPass icon. */
   background-position: calc(100% - 35px) 50% !important;
   &:focus {
@@ -47,4 +46,8 @@ export const CalendarIcon = styled.button`
   background: none;
   outline: none;
   cursor: pointer;
+
+  &:disabled {
+    cursor: not-allowed;
+  }
 `;